move most logic back into bitwise, and keep decoding and encoding logic in relevant modules
